+++ CHANGES.rst
@@ -38,6 +38,8 @@ Changelog
   `#4686 <https://github.com/syslabcom/scrum/issues/4686>`_
   [pilz]
 
+- Cleanup eea.facetednavigation, now that we have solr.
+  [pilz]
 
 2.1.1 (2026-04-05)
 ------------------